Moving test functionality to its own controller (Testing.php)
